+++ apps/flutter_architecture/lib/main_staging.dart
@@ -1,4 +1,4 @@
 import 'package:flutter_architecture/app/config/app_environment.dart';
 import 'package:flutter_architecture/bootstrap.dart';
 
-Future<void> main() => bootstrap(AppEnvironment.development);
+Future<void> main() => bootstrap(AppEnvironment.staging);